% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplyr.R
\name{bd_collect}
\alias{bd_collect}
\title{Collects the results of a remote base called via \code{bdplyr()}}
\usage{
bd_collect(.lazy_tbl, billing_project_id = basedosdados::get_billing_id())
}
\arguments{
\item{.lazy_tbl}{A variable that contains a database that was previously
connected through the \code{\link[=bdplyr]{bdplyr()}} function. Tipically, it will be called
after performing the desired operations with the \code{{dplyr}} verbs.}

\item{billing_project_id}{a string containing your billing project id.
If you've run \code{\link[=set_billing_id]{set_billing_id()}} then feel free to leave this empty.}
}
\value{
A tibble.
}
\description{
After \code{\link[=bdplyr]{bdplyr()}} is used to create the remote connection, this function
allows you to collect the result of the manipulations carried out with
the {dplyr}´s verbs and thus use it in local memory completely.

Alternatively, you can also save to disk directly using \code{\link[=bd_write]{bd_write()}}
function or its derivates: \code{\link[=bd_write_csv]{bd_write_csv()}} or \code{\link[=bd_write_rds]{bd_write_rds()}}.
}
\examples{
\dontrun{



}
}
