% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplyr.R
\name{bd_write}
\alias{bd_write}
\alias{bd_write_rds}
\alias{bd_write_csv}
\title{Writes the result of operations with \code{\link[=bdplyr]{bdplyr()}} to disk}
\usage{
bd_write(.lazy_tbl, .write_fn = `?`(typed::Function()), path, ...)

bd_write_rds(.lazy_tbl, path, overwrite = FALSE, compress = FALSE, ...)

bd_write_csv(.lazy_tbl, path, overwrite = FALSE, ...)
}
\arguments{
\item{.lazy_tbl}{A variable that contains a database that was previously
connected through the \code{\link[=bdplyr]{bdplyr()}} function. Tipically, it will be called
after performing the desired operations with the \code{{dplyr}} verbs.}

\item{.write_fn}{A function capable of writing the result of a tibble to
disk. For example: \code{\link[writexl:write_xlsx]{writexl::write_xlsx()}}, \code{\link[arrow:write_feather]{arrow::write_feather()}},
\code{\link[readr:write_delim]{readr::write_tsv()}}, etc.

#' @param path String containing the path for the file to be created.
The desired folders must already exist and the file should normallt end with
the corresponding extension.}

\item{...}{Other parameters passed to the \code{.write_fn} function.}

\item{overwrite}{FALSE by default. Indicates whether the local file should be
overwritten if it already exists. Use with care.}

\item{compress}{For \code{\link[=bd_write_rds]{bd_write_rds()}} only. A logical specifying whether
saving to a named file is to use "gzip" compression, or one of "gzip",
"bzip2" or "xz" to indicate the type of compression to be used. See also:
\code{\link[=saveRDS]{saveRDS()}}.}
}
\value{
String containing the path to the created file.
}
\description{

}
\examples{
\dontrun{

 cool_db <- basedosdados::

# pensar em exemplo json


}
}
