% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dataset_search}
\alias{dataset_search}
\title{Search for a dataset}
\arguments{
\item{search_term}{keyword for search}
}
\value{

}
\description{
Search for a dataset
}
\examples{

dataset_search("agua")
dataset_search("educação")


}
