% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{Open the data set's dictionary.}
\usage{
dictionary(dataset, billing_project_id = get_billing_id(), page_size = 1e+05)
}
\arguments{
\item{dataset}{a string containing the name of a basedosdados' data set, the dictionary of which you are asking for.}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
Returns a dictionary of the data set, allowing the user to check factors' meanings.
}
\description{
Open the data set's dictionary.
}
\examples{

\dontrun{


dictionary("br_inep_censo_escolar")

dictionary("br_me_rais")

}


}
