% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzzz.R
\name{.onLoad}
\alias{.onLoad}
\title{What happens when the package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\description{
\code{.onLoad} is automatically run when the user runs \code{library(basedosdados)} for the first time in the session.
}
