% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/billing.R
\name{get_billing_id}
\alias{get_billing_id}
\title{Internal functions for project billing management}
\usage{
get_billing_id()
}
\value{
a string with the project's billing id.
}
\description{
Retrieves the project's billing Id.
}
