% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_body}
\alias{get_body}
\title{Get body of description}
\usage{
get_body(description)
}
\arguments{
\item{description, }{a character}
}
\value{
list of character
}
\description{
Get body of description
}
