% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dataset_description}
\alias{get_dataset_description}
\title{Describe a dataset}
\usage{
get_dataset_description(dataset_id = `?`(Character(1)))
}
\arguments{
\item{dataset_id}{the dataset's name}
}
\description{
Describe a dataset
}
\examples{

get_dataset_description("br_sp_alesp")


}
