% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dataset_metadata}
\alias{get_dataset_metadata}
\title{Get Metadata Dataset}
\usage{
get_dataset_metadata(
  dataset_id,
  query_project_id = "basedosdados",
  billing_project_id = get_billing_id()
)
}
\arguments{
\item{dataset_id}{dataset id available in \code{query_project_id}}

\item{query_project_id}{which project the table lives. You can change this you want to query different projects. Default is \code{"basedosdados"}}

\item{billing_project_id}{your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{invisible(NULL)}
}
\description{
Get Metadata Dataset
}
\examples{
\dontrun{
get_dataset_metadata("br_ibge_pib")
}
}
