% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_header}
\alias{get_header}
\title{Get first line of description}
\usage{
get_header(description)
}
\arguments{
\item{description, }{a character}
}
\value{
\code{character}
}
\description{
Get first line of description
}
