% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_table_columns}
\alias{get_table_columns}
\title{Get columns in a table}
\usage{
get_table_columns(dataset_id, table_id)
}
\arguments{
\item{dataset_id}{a dataset name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{br_sp_alesp}}

\item{table_id}{a table name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{deputado}}
}
\value{
A tibble describing all columns in a table
}
\description{
Get columns in a table
}
\examples{
\dontrun{
get_table_columns("br_sp_alesp", "deputado")
}
}
