% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_table_description}
\alias{get_table_description}
\title{Describe a table within a dataset}
\usage{
get_table_description(
  dataset_id = `?`(Character(1)),
  table_id = `?`(Character(1))
)
}
\description{
Describe a table within a dataset
}
\examples{

get_table_description("br_sp_alesp", "assessores")

}
