% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_table_metadata}
\alias{get_table_metadata}
\title{Get Metadata Table}
\usage{
get_table_metadata(
  dataset_id,
  table_id,
  query_project_id = "basedosdados",
  billing_project_id = get_billing_id()
)
}
\arguments{
\item{dataset_id, }{a character, dataset id}

\item{table_id, }{a character, table id}

\item{query_project_id, }{a character, which project the table lives. You can change this you want to query different projects. Default is \code{"basedosdados"}}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{invisible(NULL)}
}
\description{
Get Metadata Table
}
\examples{
\dontrun{
get_table_metadata("br_ibge_pib", "municipios")
}
}
