% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_dataset_tables}
\alias{list_dataset_tables}
\title{List all tables in a dataset}
\usage{
list_dataset_tables(dataset_id)
}
\arguments{
\item{dataset_id}{an unique id for a dataset, such as \code{br_sp_alesp}.}
}
\description{
List all tables in a dataset
}
