% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_datasets}
\alias{list_datasets}
\title{List Datasets}
\usage{
list_datasets(
  query_project_id = "basedosdados",
  filter_by = NULL,
  with_description = FALSE,
  page_size = 100,
  ...
)
}
\arguments{
\item{query_project_id, }{Which project the table lives. You can change this you want to query different projects. Default is \code{"basedosdados"}}

\item{filter_by, }{optional character to be matched in dataset_id. Default is \code{NULL}}

\item{with_description, }{optional, if \code{TRUE}, fetch short dataset description for each dataset. Defalut is \code{FALSE}}

\item{page_size, }{Number of items per page. Default is \code{100}.}

\item{..., }{other parameters for passing in the bigrquery}
}
\value{
\code{invisible(NULL)}
}
\description{
List Datasets
}
\examples{
\dontrun{
list_datasets()
}
}
