% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table.R
\name{query_table}
\alias{query_table}
\title{Highly opinionated approach to querying - not exposed to end users, only internal}
\usage{
query_table(
  ...,
  table1,
  table2 = NULL,
  join = NULL,
  keys = NULL,
  page_size = 500,
  billing_project_id = get_billing_id()
)
}
\arguments{
\item{...}{comma-separated pairs of \verb{<table_name>.<variable_name>}}

\item{table1}{string containing the main table to be queried}

\item{table2}{string containing an optional table to be joined}

\item{join}{a join specification. Acceptable values are \code{left}, \code{right}, \code{inner} and \code{full}.}

\item{keys}{a 2-length vector containing strings with names of, respectively, \code{table1}'s foreign key and \code{table2}'s foreign key.}

\item{page_size}{\code{bigrquery} internal.}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\description{
Highly opinionated approach to querying - not exposed to end users, only internal
}
\examples{


\dontrun{
query_table(
  pib.id_municipio,
  pib.PIB,
  table1 = "basedosdados.br_ibge_pib.municipios as pib")

  }

}
