% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_table}
\alias{read_table}
\title{Read a table by name}
\usage{
read_table(table, page_size = 10000, billing_id = get_billing_id())
}
\arguments{
\item{table}{a table name}

\item{page_size}{a page_size parameter}
}
\value{
The table's content in a tibble
}
\description{
Read a table by name
}
