% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{render_dom}
\alias{render_dom}
\title{Build HTML DOM}
\usage{
render_dom(body)
}
\arguments{
\item{body}{html tree}
}
\value{
\code{invisible(NULL)}
}
\description{
Create an html document, save it in the session folder \code{tempdir()}
and open in the viewer
}
