% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{view_meta_dataset}
\alias{view_meta_dataset}
\title{View Metadata Dataset}
\usage{
view_meta_dataset(
  dataset_id,
  query_project_id = "basedosdados",
  billing_project_id = get_billing_id()
)
}
\arguments{
\item{dataset_id, }{a character, dataset id}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{NULL}
}
\description{
View Metadata Dataset
}
\examples{
\dotrun{
view_metadata_dataset("br_denatran_frota")
}
}
