% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{view_meta_table}
\alias{view_meta_table}
\title{View Metadata table}
\usage{
view_meta_table(
  dataset_id,
  table_id,
  query_project_id = "basedosdados",
  billing_project_id = get_billing_id()
)
}
\arguments{
\item{dataset_id, }{a character, dataset id}

\item{table_id, }{a character, table id}

\item{query_project_id, }{a character, which project the table lives. You can change this you want to query different projects. Default is "basedosdados"}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{NULL}
}
\description{
View Metadata table
}
\examples{
\dotrun{
view_meta_table("br_denatran_frota.municipio_tipo")
}
}
