% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{view_metadata_dataset}
\alias{view_metadata_dataset}
\title{View Metadata Dataset}
\usage{
view_metadata_dataset(dataset, billing_project_id = get_billing_id())
}
\arguments{
\item{dataset, }{a character, dataset id}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{NULL}
}
\description{
View Metadata Dataset
}
\examples{
\dotrun{
view_metadata_dataset("br_denatran_frota")
}
}
