% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{view_metadata_table}
\alias{view_metadata_table}
\title{View Metadata table}
\usage{
view_metadata_table(table, billing_project_id = get_billing_id())
}
\arguments{
\item{table, }{a character, table id}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
\code{NULL}
}
\description{
View Metadata table
}
\examples{
\dotrun{
view_metadata_table("br_denatran_frota.municipio_tipo")
}
}
