% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdplyr.R
\name{bd_write}
\alias{bd_write}
\alias{bd_write_rds}
\alias{bd_write_csv}
\title{Salva em disco o resultado de operações com bases remotas}
\usage{
bd_write(.lazy_tbl, .write_fn = `?`(typed::Function()), ...)

bd_write_rds(.lazy_tbl, path, overwrite = FALSE, compress = FALSE, ...)

bd_write_csv(.lazy_tbl, path, overwrite = FALSE, ...)
}
\arguments{
\item{.lazy_tbl}{Uma variável que foi coletada anteriormente por meio da
função \code{\link[=bdplyr]{bdplyr()}}. Recomenda-se o uso após realizadas as operações desejadas
com os verbos do pacote \code{{dplyr}}.}

\item{.write_fn}{A função de escrita desejada sem os ()}

\item{...}{Outros parâmetros que possam ser desejados.}

\item{path}{String contendo o endereço para o arquivo a ser criado. As
pastas desejadas já devem existir e o arquivo deve terminar com a extensão
correspondente.}

\item{overwrite}{Por padrão FALSE. Indica se o arquivo local deve ser
sobrescrito caso já existe. Use com cuidado.}

\item{compress}{For \code{\link[=bd_write_rds]{bd_write_rds()}} only. A logical specifying whether
saving to a named file is to use "gzip" compression, or one of "gzip",
"bzip2" or "xz" to indicate the type of compression to be used. See also:
\code{\link[=saveRDS]{saveRDS()}}.}
}
\value{
String com o endereço do arquivo salvo.
}
\description{
bla bla bla bla bla;
bla blabl bla.
}
\examples{
\dontrun{

# exemplo com csv e rds na normal e depois na helper

# pensar em exemplo json


}
}
