% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{dataset_search}
\alias{dataset_search}
\title{Search for a dataset by keyword}
\usage{
dataset_search(search_term)
}
\arguments{
\item{search_term}{keyword for search}
}
\value{
A tibble with search results
}
\description{
Search for a dataset by keyword
}
\examples{

\dontrun{

dataset_search("agua")
dataset_search("educação")

}


}
