% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_dataset_description}
\alias{get_dataset_description}
\title{Describe a dataset}
\usage{
get_dataset_description(dataset_id)
}
\arguments{
\item{dataset_id}{a dataset name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{br_sp_alesp}}
}
\value{
A tibble describing the specified dataset
}
\description{
Describe a dataset
}
\examples{

\dontrun{

get_dataset_description("br_sp_alesp")
}
}
