% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_table_description}
\alias{get_table_description}
\title{Describe a table within a dataset}
\usage{
get_table_description(
  dataset_id = `?`(typed::Character(1)),
  table_id = `?`(typed::Character(1))
)
}
\arguments{
\item{dataset_id}{a dataset name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{br_sp_alesp}}

\item{table_id}{a table name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{deputado}}
}
\value{
A tibble describing the specified table
}
\description{
Describe a table within a dataset
}
\examples{

get_table_description("br_sp_alesp", "deputado")

}
