% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_dataset_tables}
\alias{list_dataset_tables}
\title{List Dataset Tables}
\usage{
list_dataset_tables(
  dataset_id,
  query_project_id = "basedosdados",
  filter_by = NULL,
  with_description = FALSE,
  page_size = 50,
  ...
)
}
\arguments{
\item{dataset_id, }{a character, dataset id}

\item{query_project_id, }{which project the table lives. You can change this you want to query different projects. Default is \code{"basedosdados"}}

\item{filter_by, }{optional character to be matched in dataset_id. Default is \code{NULL}}

\item{with_description, }{optional, if \code{TRUE}, fetch short dataset description for each dataset. Defalut is \code{FALSE}}

\item{page_size, }{Number of items per page. Default is \code{50}.}
}
\description{
List Dataset Tables
}
