% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{list_dataset_tables}
\alias{list_dataset_tables}
\title{List tables in a dataset}
\usage{
list_dataset_tables(dataset_id)
}
\arguments{
\item{dataset_id}{a dataset name e.g. if addressing table "br_sp_alesp.deputado" then table_id is \code{br_sp_alesp}}
}
\value{
A tibble listing all tables in a given dataset
}
\description{
List tables in a dataset
}
\examples{
list_dataset_tables("br_sp_alesp")
}
