% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition_table}
\alias{partition_table}
\title{Slice a big data frame into smaller csv files by grouping variables
Still in development}
\usage{
partition_table(.data, dir, ...)
}
\arguments{
\item{.data}{a tibble.}

\item{dir}{directory where to write the csv files. Must exist before function call.}

\item{...}{comma-separated variables used to define groupings.}
}
\value{
invisibly returns all written files' addresses.
}
\description{
\code{partition_table} populates a folder
}
\examples{

\dontrun{

tibble(
  x = rnorm(1000),
  y = runif(1000) + x,
  group = sample(letters, 1000, replace = TRUE)) \%>\%
  partition_table(tempdir())


}


}
