% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{read_sql}
\alias{read_sql}
\title{Query our datalake and get results in a tibble}
\usage{
read_sql(query, billing_project_id = get_billing_id())
}
\arguments{
\item{query}{a string containing a valid SQL query.}

\item{billing_project_id}{a string containing your billing project id. If you've run \code{set_billing_id} then feel free to leave this empty.}
}
\value{
A tibble containing the query's output.
}
\description{
\code{read_sql} is given either a fully-written SQL query through the \code{query} argument or a valid table name through the \code{table} argument.
}
\examples{

\dontrun{

set_billing_id("<your id here>")

query <- "SELECT
pib.id_municipio,
pop.ano,
pib.PIB / pop.populacao * 1000 as pib_per_capita
FROM `basedosdados.br_ibge_pib.municipio` as pib
JOIN `basedosdados.br_ibge_populacao.municipio` as pop
ON pib.id_municipio = pop.id_municipio
LIMIT 5 "

data <- read_sql(query)

# in case you want to write your data on disk as a .xlsx, .csv or .Rds file.

library(writexl)
library(readr)

dir <- tempdir()

write_xlsx(data, file.path(dir, "data.xlsx"))
write_csv(data, file.path(dir, "data.csv"))
saveRDS(data, file.path(dir, "data.Rds"))

}

}
